Sampada_S_Gowda/LEGAL_POINT_GENERATOR/app/requirements.txt
google-generativeai==0.6.0
python-dotenv==1.0.0
PyPDF2==3.0.1
Pillow==9.0.0
pytesseract==0.3.10
